feat: renames ButtonCheckbox to OptionTile
